move formatToReadableDate out of utils + add missing translation
