+++ src/utilities/index.ts
@@ -12,4 +12,3 @@ export {
 export { default as calculateCollateralValue } from './calculateCollateralValue';
 export * from './generateBscScanUrl';
 export { default as formatI18nextRelativetimeValues } from './formatI18nextRelativetimeValues';
-export { default as formatToReadableDate } from './formatToReadableDate';
